# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Campus School ERP API
  description: Combined API specifications for the school ERP frontend, grouped by business domain.
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API Base Proxy Route
...
paths:
...
  /clipboard:
    get:
      summary: List Clipboard Items
      description: Fetches a list of clipboards. Supports single fetches via query.
      tags:
        - Clipboard
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Clipboard list or single item fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClipboardListResponseEnvelope'
    post:
      summary: Create a Clipboard Item
      tags:
        - Clipboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClipboardInput'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClipboardResponseEnvelope'
...
components:
...
  schemas:
...
    ClipboardInput:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          example: Project Idea Notes
        content:
          type: string
          example: Research Rich-Text framework
    ClipboardItem:
      allOf:
        - $ref: '#/components/schemas/ClipboardInput'
        - type: object
          properties:
            id:
              type: string
              example: clip-123
            created_at:
              type: string
              example: 2026-05-21
    ClipboardResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          $ref: '#/components/schemas/ClipboardItem'
    ClipboardListResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ClipboardItem'